requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
flask
pillow